pyproject.toml
[tool.poetry]
name = "prosept-team2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
django-filter = "^23.4"
gunicorn = "^21.2.0"
drf-spectacular = "^0.26.5"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
ruff = "^0.1.6"
isort = "^5.12.0"
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py311']
exclude = 'migrations'
line_length = 79

[tool.isort]
line_length = 79
skip = ['migrations', 'settings']
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true

[tool.ruff]
target-version = "py311"
exclude = [
    "*migrations/",
    "settings.py",
]
line-length = 79